despliegue de ambientes

See merge request squad-mios-v2/mios-crm-back-v2!202

+++ app/Http/Controllers/GroupController.php
@@ -41,11 +41,10 @@ class GroupController extends Controller
      */
     public function groupslist(Request $request)
     {
-        $userId = auth()->user()->rrhh_id;
-        $userLocal = User::where('id_rhh','=',$userId)->firstOrFail();
+        $rrhh_id = auth()->user()->rrhh_id;
         $groups =  Group::select('groups.id', 'groups.name_group', 'groups.description', 'groups.state')
             ->join('group_users', 'group_users.group_id', 'groups.id')
-            ->where('group_users.User_id', $userLocal->id);
+            ->where('group_users.rrhh_id', $rrhh_id);
 
         if(!is_null($request->campaign_id) && $request->campaign_id != "null")
         {
@@ -65,17 +64,20 @@ class GroupController extends Controller
     {
         try {
             $groups = new Group([
-                'campaign_id' => $request->input('campaign_id'),
+                'campaign_id' => auth()->user()->rrhh->campaign_id,
                 'name_group' => $request->input('name_group'),
                 'description' => $request->input('description'),
-                'state' => $request->input('state')
+                'state' => $request->input('state'),
+                'rrhh_id_creator' => auth()->user()->rrhh_id,
             ]);
             $groups->save();
 
-            foreach ($request->users as $userId) {
+            $users = $request->users;
+            array_push($users, ['idRrhh' => auth()->user()->rrhh_id]);
+            foreach ($users as $user) {
                 $groupsusers = new GroupUser([
                     'group_id' => $groups->id,
-                    'user_id' => $userId['userId']
+                    'rrhh_id' => $user['idRrhh']
                 ]);
                 $groupsusers->save();
             }
@@ -94,7 +96,6 @@ class GroupController extends Controller
 
     public function updateGroup(Request $request, $id)
     {
-        try {
             $groups = Group::find($id);
             $groups->name_group = $request->name_group;
             $groups->description = $request->description;
@@ -103,20 +104,21 @@ class GroupController extends Controller
 
             $groupsusers = GroupUser::where('group_id', $groups->id)->get();
             foreach ($groupsusers as $groupuser) {
-                $groupuser->delete();
+                if($groupuser->rrhh_id != $groups->rrhh_id_creator)
+                {
+                    $groupuser->delete();
+                }
             }
-            foreach ($request->users as $userId) {
+            foreach ($request->users as $user) {
                 $groupsus = new GroupUser([
                     'group_id' => $groups->id,
-                    'user_id' => $userId['userId']
+                    'rrhh_id' => $user['idRrhh']
                 ]);
                 $groupsus->save();
             }
 
             return $this->successResponse('Grupo editado Correctamente');
-        } catch (\Throwable $e) {
-            return $this->errorResponse('Error al editar el grupo', 500);
-        }
+
     }
 
     /**
@@ -128,6 +130,12 @@ class GroupController extends Controller
     {
         try {
             $group = Group::find($id);
+            $formsActive = $group->forms()->where('state', 1)->get();
+            
+            if(count($formsActive) > 0 && $request->state === 0)
+            {
+                return $this->errorResponse('El grupo no puede ser excluido por que existen formularios activos.', 500);
+            }
             $group->state = $request->state;
             $group->save();
 
@@ -144,61 +152,43 @@ class GroupController extends Controller
      */
     public function searchGroup($id)
     {
-        $idCampaign = Group::select('groups.campaign_id')->where('groups.id', $id)->firstOrFail($id)->campaign_id;
-        $usersRhh = $this->rrhhService->fetchUsersByCampaign($idCampaign);
-        $idsRrhh = $this->getIdsRrhhUsers($usersRhh);
+        //trae el id de la campanha del grupo
+        $group = Group::find($id);
 
-        $queryUsersMembers = GroupUser::join('groups', 'group_users.group_id', '=', 'groups.id')
-            ->join('users', 'group_users.user_id', '=', 'users.id')
-            ->where('groups.id', $id)
-            ->select('name_group', 'groups.description', 'group_users.user_id',
-                'users.id_rhh', 'group_users.id', 'group_users.group_id', 'groups.campaign_id');
-        $usersMembers = $queryUsersMembers->get();
+        //trae los usuarios de la camapaña
+        $usersRhh = $this->rrhhService->fetchUsersByCampaign($group->campaign_id);
 
-        $usersAvailable = User::select('users.id', 'users.id_rhh')
-            ->leftJoinSub($queryUsersMembers, 'group_users', function ($join)
-            {
-                $join->on('users.id', 'group_users.user_id');
-            })
-            ->where('group_users.id', null)
-            ->whereIn('users.id_rhh', $idsRrhh)
-            ->get();
+        //trae los usuarios del grupo
+        $usersMembersGroup = GroupUser::where('group_id', $id)->select('rrhh_id')->get();
+        $miosHelper = new MiosHelper();
+        $idsRrhhMembersGroup = $miosHelper->getArrayValues('rrhh_id', $usersMembersGroup);
 
-        $usersMembers = $this->mergeUserCrmWithUserRrhh($usersMembers, $usersRhh, true);
-        $usersAvailable = $this->mergeUserCrmWithUserRrhh($usersAvailable, $usersRhh, false);
-        $data = ['available' => $usersAvailable, 'members' => $usersMembers];
-        return $data;
+        return $this->getUserRrhhGroupMembers($idsRrhhMembersGroup, $usersRhh, $group->rrhh_id_creator);
     }
 
-    private function mergeUserCrmWithUserRrhh($userscrm, $usersRhh, $removerId)
+    private function getUserRrhhGroupMembers($idsRrhhMembersGroup, $usersRhh, $creator)
     {
-        foreach ($userscrm as $usercrm)
+        $available = [];
+        $members = [];
+        foreach ($usersRhh as $userRhh)
         {
-            foreach($usersRhh as $userRhh)
+            $user = [
+                "id_rhh" => $userRhh->id,
+                "name" => $userRhh->name
+            ];
+            if($userRhh->id != $creator)
             {
-                if($userRhh->id == $usercrm->id_rhh)
+                if(in_array($userRhh->id, $idsRrhhMembersGroup))
                 {
-                    $usercrm->name = $userRhh->name;
-                    unset($usercrm->group_id);
-                    unset($usercrm->campaign_id);
-                    if ($removerId)
-                    {
-                        unset($usercrm->id);
-                    }
+                    array_push($members, $user);
+                }
+                else
+                {
+                    array_push($available, $user);
                 }
             }
         }
-        return $userscrm;
-    }
-
-    private function getIdsRrhhUsers($usersRhh)
-    {
-        $idsRrhh = array();
-        foreach ($usersRhh as $userRhh)
-        {
-            array_push($idsRrhh, $userRhh->id);
-        }
-        return $idsRrhh;
+        return ['available' => $available, 'members' => $members];;
     }
 
     /**
@@ -209,18 +199,21 @@ class GroupController extends Controller
 
     public function searchUser($id)
     {
-        $users = $this->rrhhService->fetchUsersByCampaign($id);
-        $users = collect($users);
-        $users = $users->filter(function ($value, $key) {
-            return User::where('id_rhh', $value->id)->first() != null;
-        });
-        foreach ($users as $user) {
-            $crmUser = User::where('id_rhh', $user->id)->first();
-            if ($crmUser != null) {
-                $user->id = $crmUser->id;
+        $idCampaign = auth()->user()->rrhh->campaign_id;
+        $usersRrhh = $this->rrhhService->fetchUsersByCampaign($idCampaign);
+        $users = [];
+        foreach ($usersRrhh as $userRrhh)
+        {
+            if($userRrhh->id != auth()->user()->rrhh_id)
+            {
+                $user = [
+                    "id_rhh" => $userRrhh->id,
+                    "name" => $userRrhh->name
+                ];
+                array_push($users, $user);
             }
         }
-        return array_values($users->all());
+        return $users;
     }
 
     /**
@@ -232,7 +225,12 @@ class GroupController extends Controller
     {
         // Se obtiene los grupos por el usuario usuario 
         try {
-            $where = ['groups.state' => 1, 'group_users.user_id' => $idUser];
+            $rrhhId = auth()->user()->rrhh_id;
+            $where = ['group_users.rrhh_id' => $rrhhId];
+            if(!$this->userCanExecuteAction("ViewDisabled", "groups"))
+            {
+                $where['groups.state'] = 1;
+            }
             $groups = DB::table('groups')->join('group_users', 'groups.id', '=', 'group_users.group_id')
                 ->where($where)
                 ->select('groups.id', 'groups.campaign_id', 'groups.name_group', 'groups.description', 'groups.state', 'groups.created_at', 'groups.updated_at')
@@ -250,12 +248,12 @@ class GroupController extends Controller
      * 12-05-2021
      * Funcion para obtener las campaign que pertenecen a los grupos del usuario por id de usuario
      */
-    public function getIdCampaignByUserId($userId)
+    public function getIdCampaignByRrhhId($rrhhId)
     {
         $groups = Group::select('campaign_id')
             ->distinct()
             ->join('group_users', 'group_users.group_id', 'groups.id')
-            ->where('group_users.User_id', $userId)->get();
+            ->where('group_users.rrhh_id', $rrhhId)->get();
         $groupsIds = [];
         foreach ($groups as $group) {
             array_push($groupsIds, $group['campaign_id']);
@@ -265,7 +263,6 @@ class GroupController extends Controller
 
     public function getGroupsByRrhhId($rrhhId)
     {
-        return GroupUser::join("users", 'group_users.user_id', 'users.id')
-            ->where('users.id_rhh', $rrhhId)->get();
+        return GroupUser::where('rrhh_id', $rrhhId)->get();
     }
 }
